Fix Sidekick tab contracts and split runtime modules
